Absolute selectivity parameters
Added a slot "isRel" (character) to Fleet and OM objects.  This defaults
to TRUE, and all selectivity parameters are relative to
size-of-maturity. If set to FALSE (or false) the selectivity parameters
are absolute (e.g in mm).
Also modified the runMSE function to check if isRel is true or false,
and set the selectivity parameters appropriately.