Hack for simulated CAL and CAA data
For some reason the Grey Seal (numbers model) creates an issues where the arrays of old CAA and CAL observations have too many years and don't overwrite correctly ('incorrect dimensions' error) when simulating current data.

I've robustified this without actually diagnosing the issue. Not sure what is going on and putting a issue on the website.