Update MPs_Input.R
The older version of the DLMtool I was using had LBSPR affect effort recursively (as the paper also suggests, even though the old code was a bang-bang controller of sorts).
It seems like the new version doesn't accumulate delta efforts (V_t in the paper) so that in practice LBSPR's TAE really bounds effort only 5-10% of the time (usually on random spikes). I think without  recursivness LBSPR is basically a curE policy.